Crate cargo_util

Miscellaneous support code used by Cargo.

Modules

paths

Various utilities for working with files and paths.

registry

Structs

ProcessBuilder

A builder object for an external process, similar to std::process::Command.

ProcessError

Sha256

Functions

exit_status_to_string

Converts an ExitStatus to a human-readable string suitable for displaying to a user.

is_ci

Whether or not this running in a Continuous Integration environment.

is_simple_exit_code

Returns true if the given process exit code is something a normal process would exit with.

read2